refactoring: single source of truth is cogchain
