+++ mermaid/info.json
@@ -9,7 +9,7 @@
         "Tin Nguyen"
     ],
     "required_cogs": {},
-    "requirements": ["jinja2==3.1.6", "playwright>=1.49.0"],
+    "requirements": ["jinja2==3.1.6", "playwright>=1.49.0", "cogchain"],
     "tags": [
         "mermaid",
         "playwright",
