+++ ollama/info.json
@@ -10,7 +10,8 @@
     ],
     "required_cogs": {},
     "requirements": [
-        "langchain-ollama"
+        "langchain-ollama",
+        "cogchain"
     ],
     "tags": [
         "ollama",
