+++ qdrant/qdrant.py
@@ -9,6 +9,7 @@ from redbot.core import commands
 from redbot.core.bot import Red
 from redbot.core.config import Config
 from redbot.core.utils.chat_formatting import box
+from cogchain.interfaces import ChainStore
 
 try:
     from qdrant_client import QdrantClient
@@ -65,7 +66,7 @@ class qdrant(commands.Cog):
         )
         self._client: Optional[QdrantClient] = None
         self._collection_cache: Dict[str, Dict[str, Any]] = {}
-        self.chain_store_provider = None
+        self.chain_store_provider: Optional[ChainStore] = None
 
     async def red_delete_data_for_user(self, *, requester: RequestType, user_id: int) -> None:
         # No user-specific data is stored locally by this cog.
@@ -352,13 +353,7 @@ class qdrant(commands.Cog):
 
         return results
 
-    def _build_chain_store_provider(self):
-        try:
-            from langcore.abc import ChainStore
-        except Exception as exc:
-            log.warning("Could not import ChainStore from langcore: %s", exc)
-            return None
-
+    def _build_chain_store_provider(self) -> Optional[ChainStore]:
         cog = self
 
         class _QdrantChainStore(ChainStore):
